initial dev of xray data structures and converters
