--- exporter/awsxrayexporter/otel2xray/converter.go
@@ -1,29 +1,0 @@
-// Copyright 2019, OpenTelemetry Authors
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package otel2xray
-
-import (
-	"github.com/aws/aws-sdk-go/service/xray"
-	"github.com/open-telemetry/opentelemetry-collector/consumer/consumerdata"
-)
-
-func ConvertTraceDataToXRay(td consumerdata.TraceData) *xray.PutTraceSegmentsInput {
-	documents := make([]*string, len(td.Spans))
-	//for i, span := range td.Spans {
-	//	seg := Segment{}
-	//}
-	input := xray.PutTraceSegmentsInput{TraceSegmentDocuments: documents}
-	return &input
-}
