+++ exporter/awsxrayexporter/translator/sql.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package otel2xray
+package translator
 
 const (
 	DbTypeAttribute      = "db.type"
