+++ exporter/awsxrayexporter/translator/sql_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package otel2xray
+package translator
 
 import (
 	"github.com/stretchr/testify/assert"
